deps/v8/test/inspector/debugger/destructuring-expected.txt

Tests breakable locations in destructuring.

Running test: testBreakLocations

function testFunction() {
  function func() {
    |_|return [1, 2];|R|
  }

  var [|_|a, |_|b] = |C|func();
|R|}


Running test: testSetBreakpoint
Setting breakpoint at test.js:6:0

  var [a, b] = #func();
}

Setting breakpoint at test.js:6:7

  var [#a, b] = func();
}

Setting breakpoint at test.js:6:10

  var [a, #b] = func();
}

Setting breakpoint at test.js:6:15

  var [a, b] = #func();
}
